systemd: make emacsd start with wm@.target
Previously, emacsd started with console.target. But because of a
long-standing emacs bug that causes crashes of the daemon when X is
killed, we don't really want it to be attached to the console target:
the daemon dies without X around anywho.

So link it to my X session, and let it be done with.